ci: add .gitattributes to enforce LF for *.go files
go and especially gofmt don't play nice with CRLF, so for windows based development we need a solution. One way would be to mark everything as binary to prevent git converting to CRLF and then hope that no one checks in CRLF. Here instead I chose to force LF for all *.go files, which hopefully is more robust.